+++ PGPatcher/include/ParallaxGenConfig.hpp
@@ -124,10 +124,12 @@ public:
         struct PostPatcher {
             bool optimizeMeshes = false;
             bool fixSSS = false;
+            bool hairFlowMap = false;
 
             auto operator==(const PostPatcher& other) const -> bool
             {
-                return optimizeMeshes == other.optimizeMeshes && fixSSS == other.fixSSS;
+                return optimizeMeshes == other.optimizeMeshes && fixSSS == other.fixSSS
+                    && hairFlowMap == other.hairFlowMap;
             }
         } PostPatcher;